[iOS] Add detail text item
This CL creates a DetailTextItem for the TableView. This is a
replacement for the SettingTextItem used in CollectionView.

Here is a screenshot on the Table Cell Catalog, the cell with "Item with
two labels" and "The second label..."
https://drive.google.com/file/d/1cgTIs8h7mR8Ae8rnYSjvYZu9sMeqwxmX/view

Bug: 894791
Cq-Include-Trybots: luci.chromium.try:ios-simulator-cronet;luci.chromium.try:ios-simulator-full-configs
Change-Id: Iab73928fca89131f326c7bbfc3c2bbb16fff5442
Reviewed-on: https://chromium-review.googlesource.com/c/1286141
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#601103}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9962b26e3b416e540f1f8c2d40369b3f99774cd