[Autofill] Add strikes when CC save ignored/denied/fails
This is a retry/relanding of
https://chromium-review.googlesource.com/c/chromium/src/+/1269524,
but with the browsertest additions removed, because something is causing
them to be flaky on Win7 builds.  Despite making progress, the original
CL has already been reverted twice, so I'd like to land this portion of
it so if it's reverted in the future, only the browsertests (done as a
subsequent CL) will be reverted instead of this big 18-file thing.

Bug: 884817
Change-Id: I05983ff8b13c7dbac25e615bba8a5d1fa7cc0819
Reviewed-on: https://chromium-review.googlesource.com/c/1294429
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Jared Saul <jsaul@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#601720}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb15c8040b58a4e7d9d3adf7729ebf9904d82e28