[iOS] Change the layout of FindBar
This CL changes the layout of the FindBar so it is displayed on top of
the toolbar on iPhone. Previously it was displayed below the safe area
and with a fixed height. Now it is covering the toolbar.

Bug: 892145
Change-Id: I06aac98ad9396373a78bbd3fb4fb8691e6a2a82b
Reviewed-on: https://chromium-review.googlesource.com/c/1290971
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#602715}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a056c55765a5def12a8c55c592a16f45fd9ca42