[ios] Sets current BrowserState to RecentTabsCoordinator.
Since the OpenTab commands needs to know if the Tab should be open in
Incognito or not, we should send the current BrowserState to the VC
so it gets this information.

Conversely all other services depend on the OriginalBrowserState, so we
make sure to use this one for those.

Bug: 895341
Change-Id: Idd7439a81432ec87d201e76d7fd60abec2615bbf
Reviewed-on: https://chromium-review.googlesource.com/c/1299928
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#603978}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 067cd01621e0ed2489b5a346481b6c9fbeeed59b