Add (a no-op) loadPage action to iOS autofill automation.
The Autofill Recorder Extension records an action when a new page is
loaded as part of a button/link click, etc (in addition to the action
that performs that click).

I found that desktop no-ops this action, and that on iOS it also does
not need to be explicitly replayed, so I no-op it here as well. I am
discussing with Yiming the reason it is no-op'd on desktop, and will
add that information here when I receive it.

Change-Id: If5a2f49cf60995831dfae357c97de28266771ea2
Reviewed-on: https://chromium-review.googlesource.com/c/1307994
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: ericale <ericale@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#604061}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7433188782a4c2b192c5cff26b815ccd7c512563