Add a flag needed for Autofill tests to test_runner.py.
This was configured in my scheme when testing locally, so I forgot it
even existed. Found when doing a full repro of the bot testing flow
while trying to debug unrelated issues with ios-wpr-simulator.

This line is very long as both of these flags should be part of "-c"
so that iossim passes it properly. Is that fine?

Bug: 892381
Change-Id: I5d7def3f138d10033914718429307b76a8d71d8a
Reviewed-on: https://chromium-review.googlesource.com/c/1311809
Commit-Queue: ericale <ericale@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#604615}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: faf94d6de2f91071d7ddb76f245bd8643acac3fd