[ios] Removes Layout() InfobarIOS method.
Since the ContainerView size is now dependent on the InfobarView
(The ContainerView autolayout constraints are anchored against the
InfobarView, and we set the containerView to match the InfobarView height),
there's no need to pass the ContainerView frame to layout the InfobarView.

Because of this InfobarContainerVC no longer needs to import InfobarIOS
and its moved to the infobars_ui BUILD.gn source_set.

Bug: 892376
Change-Id: I44ca8c470b0ededbd34fe2f1564c2d651f200317
Reviewed-on: https://chromium-review.googlesource.com/c/1315987
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#606504}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fbae555786d079903f5b131966682ab8d260526