[iOS][AF] Pass suggestions instead of UIViews
Modifies the pipeline so suggestions_controller responds with a model
for suggestions instead of a customized UIView. This gives the accessory
bar the responsibility (and the power) of creating the view.

This is the first step to allow better animations, by not having
to completely replace those views every time new suggestions arrive. As
well as having an easy way to track the number of suggestions when a
manual fallback icon is tapped.

Bug: 8445472
Change-Id: I0a58868de27e37af98a15d913782ddbc7967bbae
Reviewed-on: https://chromium-review.googlesource.com/c/1327162
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#607564}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de0008c49b2b858460b3220bb537794e1803fdc1