[iOS] Adjust pinnedYOffset value to neareset pixel
This CL makes sure that the value returned by -pinnedYOffset is adjusted
to the nearest pixel. This is because when the content size isn't the
default one, the value isn't adjusted on a pixel. When it is used to
compare to the scroll value of the collection view, this gap prevents
the collection view from being considered as completely scrolled.

Bug: 905276
Change-Id: Ia44de6063b039a022fcde95e015c3e7df4259308
Reviewed-on: https://chromium-review.googlesource.com/c/1337490
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#608748}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87899a86f424f3d67be019d13d9e70f853a7a62f