Do not start SadTabCoordinator for invisible Tab.
This fixes the issue when side swiping from one SadTab to another.
Since -[SadTabCoordinator start] is idempotent it is important to start
coordinator for new sad tab, not for the old one which is not visible
anymore.

Bug: 901563
Change-Id: I9dfb488037da18be5fdb1079e849b723fa723c08
Reviewed-on: https://chromium-review.googlesource.com/c/1337814
Commit-Queue: Eugene But <eugenebut@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#609011}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 99162f9e29b36c069d244c6f5dc432674504085b