[iOS] Change the NTP animation of the location bar
This CL changes the way to NTP is animating the appearance of the
location bar in the toolbar when the device has a regular width.
It uses a transform instead of changing the bottom constraint of the
location bar.

Bug: 906176
Change-Id: Ia110cafd9d2093a82487b2e3c342337c8cf7ef12
Reviewed-on: https://chromium-review.googlesource.com/c/1341911
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#609340}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a289922bbc39f8d70ac7abb580b300f2e330ad1c