[Device info sync] Construct the controller in ProfileSyncService
This CL moves the construction of the sync controller for device info
directly into ProfileSyncService that owns the DeviceInfoSyncBridge.
This (a) heavily simplifies how the controller gets access to its
delegate (removing the accessor from SyncClients) and (b) allows to use
the forwarding delegate instead of the proxy delegate.

This CL is a preparation for using the simpler forwarding delegate for
all sync types that run on the UI sequence.

This CL also resolves a TODO by making it clear that DEVICE_INFO cannot
get disabled via the command-line switch.

Bug: 867801
Change-Id: Iea84d818613476f40d89eebd73266aa24d028fb6
Reviewed-on: https://chromium-review.googlesource.com/c/1348092
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#610448}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4e2ea188dc6fdb21c332d32f7cfc870202955015