[iOS] Don't prevent toolbar contract while animating
This CL let the toolbar be contracted even if it is animating.
This is needed, because calling twice the same method could yield
different results based on the state of the device (portrait vs
landscape for example).

Bug: 900761
Change-Id: Ie3f44e871e411162bc63467d6cf5e7df3d02cc22
Reviewed-on: https://chromium-review.googlesource.com/c/1349347
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#611077}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90bd4f445a5689271c9e3e9e95021f630418b9a8