Marked NavigationItemsPruned as deprecated.
DidChangeBackForwardState is a superset of this callback and should be
used instead of NavigationItemsPruned in the future.
NavigationItemsPruned is not called when slim-navigation-manager is
enabled and DidChangeBackForwardState is not called when
slim-navigation-manager is disabled. So for now the clients should
implement both callbacks.

Bug: 910894
Change-Id: I39a3972a909e956b92e0c68a66ce346c49833be7
Reviewed-on: https://chromium-review.googlesource.com/c/1368948
Reviewed-by: Danyao Wang <danyao@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#614886}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dca91f47f7dacc724352ef9b2055b6d7740eefb0