[Autofill] Add SaveCardOfferUserDecision to save card callbacks
Instead of having CreditCardSaveManager be contacted via a callback
only whenever credit card save is accepted, this change causes it to
also be notified if the user ignored or outright rejected the credit
card save prompt.

This is required in order to implement iOS-based browsertests for the
StrikeDatabase system, but it also has the additional bonus of removing
calls to LegacyStrikeDatabase from UI code, placing them back in the
hands of CreditCardSaveManager (where they should be).

Bug: 884817
Change-Id: I84adfcf6520361613b107dedccd3849cc7a14671
Reviewed-on: https://chromium-review.googlesource.com/c/1357725
Commit-Queue: Jared Saul <jsaul@google.com>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#614925}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 565c28a3e27938b8fd4f638587b2c26895e4d547