Use define to provide a different user agent product for //ios/web_view.
This is needed to differentiate //ios/web_view from //ios/chrome as a
client of sync.

"ChromeWebView" is chosen as the product name for //ios/web_view since
that is the name of the binary. Also, when combined with IOS_PHONE,
one can reasonable deduce that this refers to //ios/web_view.

Bug: 872420
Change-Id: Ib3b849935515feca766291ed3d35e4431d49fcb1
Reviewed-on: https://chromium-review.googlesource.com/c/1371049
Commit-Queue: John Wu <jzw@chromium.org>
Reviewed-by: Hiroshi Ichikawa <ichikawa@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#617063}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 44fda8cbaeb219a7a4768bad467cb7a45a1f7928